misc/HtmlTestSuite/results/test-0885.xhtml

<?xml version="1.0" encoding="UTF-8"?><html xmlns="http://www.w3.org/1999/xhtml"><head><title>Cyclic linking between JavaScript and WebAssembly (JS higher)</title>

<script src="/resources/testharness.js"/>
<script src="/resources/testharnessreport.js"/>
<script>
    setup({allow_uncaught_exception: true});

    const test_load = async_test(
        "Check cyclic linking between JavaScript and WebAssembly where JavaScript is higher in the module graph.");

    window.log = [];
    window.addEventListener("error", ev =&gt; {
      test_load.step(() =&gt; assert_equals(ev.error.constructor, WebAssembly.LinkError));
      log.push(ev.message);
    });

    window.addEventListener("load", test_load.step_func_done(ev =&gt; {
      assert_equals(log.length, 10);
      assert_equals(log[1], 1);
      assert_equals(log[3], 2);
      assert_equals(log[5], 3);
      assert_equals(log[7], 4);
      assert_equals(log[9], 5);
    }));

    function unreachable() { log.push("unexpected"); }
</script>
<script onerror="unreachable()" onload="log.push(1)" src="./resources/js-wasm-cycle-value.js" type="module"/>
<script onerror="unreachable()" onload="log.push(2)" src="./resources/js-wasm-cycle-global.js" type="module"/>
<script onerror="unreachable()" onload="log.push(3)" src="./resources/js-wasm-cycle-memory.js" type="module"/>
<script onerror="unreachable()" onload="log.push(4)" src="./resources/js-wasm-cycle-table.js" type="module"/>
<script onerror="unreachable()" onload="log.push(5)" src="./resources/js-wasm-cycle-function-error.js" type="module"/>
</head><body/></html>
